fix: extract formatPhoneNumber to own module to fix CI coverage
The formatPhoneNumber spec imported from RegistrationCheckoutForm.tsx,
which caused V8 to instrument the entire registrations lib (including
SquareCardForm at 1% coverage), dropping global coverage from 81% to 72%.

Extracting the pure utility function to its own file keeps the test
isolated and prevents heavy React components from polluting coverage.

Also excludes .claude/ worktrees from vitest to prevent local test
pollution from stale worktree directories.
