+++ libs/react/registrations/src/lib/RegistrationCheckoutForm.tsx
@@ -16,18 +16,7 @@ import type {
   CalculateRegistrationCostResponse,
   CreateRegistrationResponse,
 } from '@maple/ts/firebase/api-types';
-
-/**
- * Formats a string of digits into US phone format: (XXX) XXX-XXXX.
- * Strips all non-digit characters first, then applies the mask progressively.
- */
-export function formatPhoneNumber(value: string): string {
-  const digits = value.replace(/\D/g, '').slice(0, 10);
-  if (digits.length === 0) return '';
-  if (digits.length <= 3) return `(${digits}`;
-  if (digits.length <= 6) return `(${digits.slice(0, 3)}) ${digits.slice(3)}`;
-  return `(${digits.slice(0, 3)}) ${digits.slice(3, 6)}-${digits.slice(6)}`;
-}
+import { formatPhoneNumber } from './formatPhoneNumber';
 
 interface RegistrationCheckoutFormProps {
   publicClass: PublicClass;
